Add solution for the online game task
The task needs following the exact instructions provided. Initially it
seemed like a knapsack problem, but this was not the case in the end.

We iterate over the sorted list of teams as long as we can fill our
group, but we do not prioritise many teams with smaller scores over a
team with a larger score, even if that would lead to a better server
joining queue.